backend/app/config.py
import os
from dotenv import load_dotenv

load_dotenv()


class Settings:
    USE_VERTEX_AI: bool = os.getenv("USE_VERTEX_AI", "true").lower() == "true"

    GOOGLE_CLOUD_PROJECT: str | None = os.getenv("GOOGLE_CLOUD_PROJECT")
    GOOGLE_CLOUD_LOCATION: str = os.getenv("GOOGLE_CLOUD_LOCATION", "global")

    GEMINI_TEXT_MODEL: str = os.getenv("GEMINI_TEXT_MODEL", "gemini-2.5-flash")
    GEMINI_EMBED_MODEL: str = os.getenv("GEMINI_EMBED_MODEL", "gemini-embedding-001")


settings = Settings()