backend/requirements.txt
streamlit
openai
pandas
matplotlib
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-dotenv==1.0.1
pydantic==2.11.7
google-genai==1.30.0
numpy==2.2.6